+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoacController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityCoverController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityParticipantController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityFileController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。